+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 
 ### Added
 
+- `api`
+  - Better inform the user when the recorder is prepared and ready to process
+    audio. We now send out an `recorderready` event if ready.
 - `examples`
   - Added the `react-speechchallenge` example. This example shows a speech
     challenge example and uses the recorder to record audio for it. It is also
@@ -16,6 +19,10 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 
 ### Changed
 
+- `api`
+  - The broadcaster event `websocketserverreadyforaudio` is a bit confusing.
+    Sending it through the event emitter is deprecated and will be removed in a
+    future version.
 - `sdk`
   - Updated .gitignore file with more common patterns.